requirements.txt
altgraph==0.17.4
astroid==3.3.5
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
fonttools==4.55.0
isort==5.13.2
kiwisolver==1.4.7
macholib==1.16.3
matplotlib==3.9.2
mccabe==0.7.0
numpy==2.1.3
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
psutil==6.1.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pylint==3.3.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
setuptools==75.2.0
six==1.16.0
speedtest-cli==2.1.3
tk==0.1.0
tomlkit==0.13.2
tqdm==4.66.6
ttkbootstrap>=1.10.1
